Ticket 49331 - change autoscaling defaults
Bug Description:  In autotuning we default to a split of 40% to
dbcache and 60 to entry. This isn't always the best split, given
we cap the dbcache.

In some performance tests I noticed an improvement when increasing
the dbcache so I want to propose we change the scaling to 25% split
and a cap of 1.5GB.

Fix Description:  Change the values to 25% and 1.5GB cap

https://pagure.io/389-ds-base/issue/49331

Author: wibrown

Review by: mreynolds (Thanks!)