Ticket 49348 - support perlless and wrapperless install
Bug Description:  We have a number of shell wrappers that now
server no purpose - they existed to support non-standard
library paths, but this issue is not true of autotools in 2017.

We also want to remove perl in favour of python. This adds a
non-default option to produce a perl-less server install, which
will force us to adapt our lib389 test suites to workout without
the perl tools.

Fix Description:  Add support for a "--disable-perl" flag to
configure. Remove the shell library wrappers, and rename the
targets in makefile. Fix an issue in task.c and main.c with
normdn and error message clarity. Finally, the major component
was improvements and polish to the basic_test suite to fully
support python 3 with a perl-less install.

See also: https://pagure.io/lib389/issue/88

https://pagure.io/389-ds-base/issue/49348

Author: wibrown

Review by: mreynolds (THanks!)