Ticket 49356 - mapping tree crash can occur during tot init
Bug Description:  Two faults were found in the handling of the mapping
tree of 389 directory server. The first fault was that the tree-free
check was not performed atomically and may cause an incorrect operations
error to be returned. The second was that during a total init the referral
would not lock the be, but the pw_verify code assumed a be was locked.
This caused a segfault.

Fix Description:  Fix the freed check to use atomics. Fix the pw_verify
to assert be is NULL (which is correct, there is no backend).

https://pagure.io/389-ds-base/issue/49356

Author: wibrown

Review by: mreynolds (THanks!)