Ticket 49392 - memavailable not available
Bug Description:  On certain linux platforms memAvailable is
not actually available! This means that the value was 0, so
cgroup max was read instead, setting the system ram to:

9223372036854771712

That's a bit excessive, and can cause memory allocations to fail.

Fix Description:  If memavail can't be found, fall back to
memtotal instead.

https://pagure.io/389-ds-base/issue/49392

Author: wibrown

Review by: mreynolds (Thanks!)