Ticket 49444 - huaf in task.c during high load import
Bug Description:  Due to incorrect locking in task.c, it was possible
for task.c to heap-use-after-free on the task-log message causing
corruption or crashing during import of ldif.

Fix Description:  Correct the locking to be after the mods are
commited to prevent the realloc from freeing before mod completes.

https://pagure.io/389-ds-base/issue/49444

Author: wibrown

Review by: lkrispen (Thank you!)