Ticket 49298 - Correct error codes with config restore.
Bug Description:  The piece of code uses 0 as an error - not 1,
and in some cases did not even check the codes or use the
correct logic.

Fix Description:  Cleanup dse_check_file to better check the
content of files and communicate issues to the admin. Correct
slapd_bootstrap_config to correctly handle the cases of removal
and restore.

https://pagure.io/389-ds-base/issue/49298

Author: wibrown

Review by: mreynoolds & spichugi

Signed-off-by: Mark Reynolds <mreynolds@redhat.com>