Ticket 49461 - Improve db2index handling for test 49290
Bug Description:  db2index has limited parameters we can accept,
including the types of attributes we can index. Specificatlly,
we must take an index name, else we do not re-index (look at
ns-slapd, we require -t or -T, and if we do, we require them
to have a value.

Fix Description:  Fix lib389 to match the assertions of ns-slapd
with regard to db2index behaviour.

https://pagure.io/389-ds-base/issue/49461

Author: wibrown

Review by: mreynolds (Thanks!)