Ticket 49325 - fix rust linking.
Bug Description:  An issue with the rpm was found that the way
the static .a was built would cause a missing library error when
the rpm was distributed.

Fix Description:  Because both cargo and automake are opinionated
and stubborn, this leaves the option as have cargo generate the .so
but write manual rules for install and linking. This has been tested
now with non-prefix, prefix, rpm build, copr.

https://pagure.io/389-ds-base/issue/49325

Author: wibrown

Review by: mreynolds (Thanks!)