Ticket 49449 - Load sysctl values on rpm upgrade.
Bug Description:  We provide sysctl values. We should apply them
on upgrade or installation to prevent a need to reboot.

Fix Description:  Add a post install script to load the sysctl
values from the system as they wolud be on a reboot. Note this
may not work in all cases (ie mock, container etc) so we have
to mask failures "just in case".

https://pagure.io/389-ds-base/issue/49449

Author: wibrown

Review by: tbordaz (Thank you!)