Ticket 49474 - purge saslmaps before gssapi test
Bug Description:  In the GSSAPI test, depending on enviroment
some of the default maps can "work" or they "may not". This is
due to inconsistency in the behaviour of the cyrus-sasl
library and it's assignment of realm into requests.

Fix Description:  Purge all maps by default, and just use the
provided test sasl map - this is a single "source of truth"
and thus allows us to test assertions easier.

https://pagure.io/389-ds-base/issue/49474

Author: wibrown

Review by: spichugi (Thank you!)