Ticket 49447 -  PBKDF2 on upgrade
Bug Description:  We changed the default password hash mech to
PBKDF2 on upgrade - but if the upgrade from 1.3.x to 1.4.x,
this new plugin object doesn't exist. This causes the server to
fail to start.

Adittionally, one could imagine an admin error could cause this
also to be omitted

Fix Description:  Create and bootstrap the plugins on startup

https://pagure.io/389-ds-base/issue/49447

Author: wibrown

Review by: mreynolds