Issue 49538 - replace cacertdir_rehash with openssl rehash
Bug description: Enable TLS method fails on newer Fedora systems.
It complains that cacertdir_rehash tool is not found.

Fix description: The reason is that authconfig was replaced with
authselect package. Authconfig shipped a tool called cacertdir_rehash
which is no longer available on Fedora. We need to switch to openssl tool:
'c_rehash <directory>' that serves the same purpose and is present on both
RHEL7 and Fedora.

Remove authconfig from 389-ds-base.spec.in because 'cacertdir_rehash'
was the only reason why we had the dependency.
Add openssl-perl dependency to the SPEC file.

Eventially, c_rehash will go away but for now we keep it for RHEL7 compatibility.

Fix a small thing in remove_ds_instance function.
We should ignore FileNotFound errors while removing the instance.

https://pagure.io/389-ds-base/issue/49538

Reviewed by: mreynolds, mhonek (Thanks!)