Ticket 49650 - lib389 enable_tls doesn't work on F28
Bug Description: In the lib389 we have the method inst.enable_tls().
It creates certificates and sets up the server for TLS communication.
It works on F27 built from master branch and doesn't work on F28.
It happens because on F28 openssl fails to verify the certificate.

Fix Description: We should create CA with an appropriate flag.
It can be done by setting an X.509 V3 Certificate Type Extension
in the certificate to 'certSigning' value.

https://pagure.io/389-ds-base/issue/49650

Author: mhonek

Review by: spichugi, wibrown (Thanks!)