Issue 49678 - organiSational vs organiZational spelling in lib389
Bug description:
RFC4519 defines 'organization' object class using Oxford spelling.
It is common for academic, formal, and technical writing for an
international readership. We should be following common spelling
in function names and methods, otherwise it is confusing
and annoying to remember which one is used where.

Fix description:
Use Oxford spelling for 'organization' and 'organizationalUnit'.

https://pagure.io/389-ds-base/issue/49678

Reviewed by: spichugi, mhonek (Thanks!)