Issue 49761 - Fix test suite issues
Description: A lot of tests are failing because of the API changes.
We need to fix them ASAP.

Fix description: Add __init__.py to import and mapping_tree test suite
so it will create its own __pycache__.
Use the new Agriments object for reinit.
Run range_search memory leak test only when ASAN is enabled and
remove valgrind support because ASAN is enough.
Generate ou=People and ou=Groups in the dbgen.py module, it is
required by some tests.
In the replica.py module, use existing credentials when we join master to
the existing first_master and  when we failed to get it from the ReplicationManager.

https://pagure.io/389-ds-base/issue/49761

Reviewed by: mreynolds (Thanks!)