Ticket 49887: Fix SASL map creation when --disable-perl
When the SASL maps for single backend was made, it was assuming a number
of behaviours such as a single backend was added. This is *not* the
default in lib389, and caused a crashed when a valid ZERO backend
server was created. Additionally, it used a template file that
will not be present in the future.

This Fixes the behaviour to account for 0, 1 and multiple backends
and uses the correct saslmapping objects to create the types needed
for LDAPI

https://pagure.io/389-ds-base/issue/49887

Author: William Brown <william@blackhats.net.au>