Ticket 49574 - remove index subsystem
There is code for an indexing subsystem, which looks like it was intended
to provide special indexes for specific search filters (not vlv).
But it is not used and not documented and unclear if it ever worked or will
work. It only contains calls in search processing to check if something is defined
and so only provides a bit of overhead in searches.

We can remove this dead code. We can also remove the broker api for idl_xxx since it is not used
anywhere else.