Ticket 50077 - RFE - improve automember plugin to work with modify ops
Description:

Previously automember was only invoked for ADD operations.  This enhancement
allows it to work with modify operations, and it will also maintain the
correct memberships.  So if a modify changes which groups the user would
belong to, it will add the user to the new group, and remove them from the
old group.

https://pagure.io/389-ds-base/issue/50077

Reviewed by: spichugi & firstyear (Thanks!!)