Issue 50122 - Selinux test for presence
Description:  Selinux is not present on all systems. Trying to
import python-selinux when it's not available fails, but we
can not guarantee that the with_selinux flag to defaults is
correct because some systems build with selinux but may not
have it enabled. We should check if we can access the tools
instead, and skip them (with warnings) if we can't make changes

https://pagure.io/389-ds-base/issue/50122

Reviewed by: mreynolds