Ticket 50126 - Incorrect usage of sudo in test
Bug Description:

Sudo is assumed to be present and working in passwordless
mode with basictest.

Fix Description:

We can not make this assumption, that sudo is installed (docker)
or in passwordless mode. As a result for this test to work, we should
run the suite as "sudo py.test" or "dirsrv". Potentially we may need
to update defaults.inf for people who want to run the tests as
their own user ID.

https://pagure.io/389-ds-base/issue/50126

Author: William Brown <william@blackhats.net.au>

Review by: ???