Ticket 50138 - db2bak.pl -P LDAPS does not work when nsslapd-securePort is missing
Bug Description:
When nsslapd-securePort is not set in dse.ldif, DSUtil (in case of LDAPS)
interpolates ldapmodify (and other) call to a URI ending only with a colon
because obtained securePort is just an empty string. The very same turns out to
happen for the nsslapd-port.

Fix Description:
Hardcode the default ports, which is a short variant of omitting the
colon and port, and which would be chosen by the OpenLDAP utils any way.

https://pagure.io/389-ds-base/issue/50138

Author: mhonek

Review by: mreynolds (thanks!)