Ticket 50140 - Use high ports in container installs
Bug Description: Out of the box, linux and containers don't
have the required root permissions to use ports below 1024.
We can't expect admins to change this, so we should configure
ourselves on high ports in container installs.

Fix Description: Add containised argument to slapd2base
options, and pass it as required for example file and
installer sections.

https://pagure.io/389-ds-base/issue/50140

Author: William Brown <william@blackhats.net.au>

Review by: ???