Ticket 49390 - improve compare and cn=config compare tests
Bug Description: We had a number of tests for the dsldapobject
compare cases, but they were in the lib389 tests. Move and update
these to work as part of the dirsrvtests suite.

Fix Description: Update lib389 to properly handle attribute casing
and update compare tests to work with newer lib389 ideas

https://pagure.io/389-ds-base/issue/49390

Author: William Brown <william@blackhats.net.au>

Review by: spichugi (Thanks)