Ticket 50310 - fix sasl header include
Bug Description: After the merge of the PKG_CONFIG change, on SUSE
the server fails to build. This is because the pkg-config for
sasl on suse doesn't add the -I include for sasl to the path so
using sasl.h doesn't work.

Fix Description: Change all references to sasl/sasl.h

https://pagure.io/389-ds-base/issue/50310

Author: William Brown <william@blackhats.net.au>

Review by: hmc, mreynolds (thanks!)