Ticket 50344 - tidy rpm vs build systemd flag handling
Bug Description: In rpm builds we would read with_systemd from
defaults.inf, which has a diffeent value to hand-building. AS
a result this caused as issue in dscontainer on opensuse where
it believed systemd was present.

Fix Description: Simplify the systemd handling to a single flag
which is possible to override in a container env.

https://pagure.io/389-ds-base/issue/50344

Author: William Brown <william@blackhats.net.au>

Review by: ???