Ticket 50439 - Update docker integration to work out of source directory
Bug Description: Docker did not function in some cases, and we had to wait for
releases via rpm.

Fix Description: This adds the support to build from source into the tree
so that we can build and test git master. This also resolves a var/run
issue in the image, as well as some other minor python cleaning such
as handling sigchld to act as init.

https://pagure.io/389-ds-base/issue/50439

Author: William Brown william@blackhats.net.au

Review by: spichugi