Ticket 50439 - fix waitpid issue when pid does not exist
Bug Description: In some situations, waitpid will fail with
a no child process error, when the pid file has a value but
no pid exists.

Fix Description: Catch the exception, because in this case
we have no pids to wait upon, so there is no harm to skip this.

https://pagure.io/389-ds-base/issue/50439

Author: William Brown <william@blackhats.net.au>

Review by: ???