Ticket 50459 - Correct issue with allocation state
Bug Description: While adding the connection state, due to
a misunderstanding on my part, it was possible that a connection
was more likely to fail to allocate causing the server to exit(1)
incorrectly.

Fix Description: Fix the state handler to correctly account for
connection structure reuse.

https://pagure.io/389-ds-base/issue/50459

Author: William Brown <william@blackhats.net.au>

Review by: tbordaz (Thanks!)