Issue 49624 - modrdn silently fails if DB deadlock occurs
Bug Description:

If a DB Deadlock error occurs during a modrdn operation the entry
cache gets updated (corrupted), but the update is not applied to
the database.

Fix Description:

Looks like there was a copy & paste error, and the wrong attribute
was updated during the retry of the modrdn operation.

relates: https://pagure.io/389-ds-base/issue/49624

Reviewed by: lkrispenz (Thanks!)