Ticket 50584, 49212 - docker healthcheck and configuration
Bug Description: Docker is managed by providing values from the
environment. To know if this is ready to make changes to our
local instance, we need to be able to check the instance is
healthy. In addition, docker has a health check process which
can allow monitoring and management of instances as they start.

Fix Description: This provides a healthcheck tool and allows
configuration by the envirnoment for the directory manager
password, and allows indicating via the env to perform a
db2index on startup.

https://pagure.io/389-ds-base/issue/49212
https://pagure.io/389-ds-base/issue/50584

Author: William Brown <william@blackhats.net.au>

Review by: ???